+++ packages/styleguide/.storybook/addons/system/preset.ts
@@ -1,4 +1,4 @@
-function managerEntries(entry: any[] = [], options: any) {
+function managerEntries(entry: any[] = []) {
   return [...entry, require.resolve('./register')];
 }